projects/paper2code-gemini-implementation/pyproject.toml
[project]
name = "paper2code-gemini"
version = "0.1.0"
description = "Paper2Code pipeline powered by the Gemini API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.118.0",
    "uvicorn>=0.37.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.11.0",
]
